NLP_LLM_Projects/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
numpy
openai